Lets make your Yoga Mat a little bit smarter!
This is Android source code for a Smart Mat project which was created
during Junction 2015 Hackathon in Helsinki. The Android Application
communicates with pressure sensors located under a Yoga Mat. When
someone sits on the mat, the mat sends BLE notification to the mobile
phone with a non-zero value. It sends another notification after the
person leaves sitting on mat with a value of zero. Time duration between
two value is used to calculate the mindfullness level. The more one
sits, the more the level is increased. Each time the application goes to
the background, the mindfullness level is decreased. This can be
extended in a way that application collects the phone usage statistics
using Android Usage API
(http://developer.android.com/reference/android/app/usage/package-summary.html)
and decreases the mindfullness level accordingly. The idea is preventing
social media addiction by doing yoga on a smart Mat.

In the Main Activity source code the UI is implemented in a really
absurd way due to limited time in Hackathon. The main idea was adding a
green button behind the cropped transparent man doing yoga and
increasing the button size level by level. However, this idea failed due
to some problems and I came up with adding 35 background images each
indicates different level. As the level increases or decreases the
background image changes :)

Currently, the Custom Mat Service should have only one BluetoothGatt
characteristic with Notification property which notifies the state 1 or
0 (Pressure or No Pressure). However tHe mobile app also looks for one
more Bluetooth LE Characteristic to write a value to the Smart Mat! The
UUID values can be found in MatProfile.java file!

As I said this was just one day hack and feel free to extend this app
with your ideas!